chore(action-bar): move slots to separate file
