+++ src/components/ActionBar/ActionBar.slots.jsx
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+import React from 'react';
+
+export const ActionBarContentSection = ({ children }) => <>{children}</>;
+
+ActionBarContentSection.propTypes = {
+	children: PropTypes.node,
+};
