+++ src/components/ActionBar/ActionBar.stories.mdx
@@ -2,7 +2,8 @@ import { Fragment } from 'react';
 import { action } from '@storybook/addon-actions';
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 
-import ActionBar, { ActionBarContentSection } from './ActionBar';
+import ActionBar from './ActionBar';
+import { ActionBarContentSection } from './ActionBar.slots';
 
 <Meta title="ActionBar" component={ActionBar} />
